PUB-1311 Remove collections page (#420)
* Interstitial page removed and changed to home page

* Added unit test

* Fix conflicts
