+++ src/main/controllers/HomeController.ts
@@ -1,8 +1,13 @@
-import { Application } from 'express';
-import {PipRequest} from '../models/request/PipRequest';
+import { PipRequest } from '../models/request/PipRequest';
+import { Response } from 'express';
+import { cloneDeep } from 'lodash';
 
 export default class HomeController {
-  public get(req: PipRequest, res: Application): void {
-    res.render('home', req.i18n.getDataByLanguage(req.lng).home);
+  public get(req: PipRequest, res: Response): void {
+    const currentLanguage = req?.lng ? req.lng : 'en';
+    res.render('home', {
+      ...cloneDeep(req.i18n.getDataByLanguage(req.lng).home),
+      currentLanguage,
+    });
   }
 }
