+++ src/main/resources/locales/cy/home.json
@@ -1,7 +1,7 @@
 {
   "title": "Gwrandawiadau llys a thribiwnlys",
   "message": "Gallwch ddefnyddio'r gwasanaeth hwn i gael gwybodaeth am:",
-  "betaHeading": "Mae hwn yn wasanaeth newydd – bydd eich <a id=\"betalink\" class=\"govuk-link\" href=\"https://www.smartsurvey.co.uk/s/FBSPI22/?pageurl=interstitial\">adborth</a> yn ein helpu i’w wella",
+  "betaHeading": "Mae hwn yn wasanaeth newydd – bydd eich <a id=\"betalink\" class=\"govuk-link\" href=\"https://www.smartsurvey.co.uk/s/FBSPI22/?pageurl=/\">adborth</a> yn ein helpu i’w wella",
   "bullets": [
     "Hearings in Civil and Family Courts in Milton Keynes, Oxford, and Reading",
     "Achosion Gweithdrefn Un Ynad, gan gynnwys trwyddedu teledu a mân droseddau traffig megis goryrru"
