--- src/main/resources/locales/en/interstitial.json
@@ -1,30 +1,0 @@
-{
-  "title": "Court and tribunal hearings",
-  "message": "You can use this service to get information about:",
-  "betaHeading": "This is a new service – your <a id=\"betalink\" class=\"govuk-link\" href=\"https://www.smartsurvey.co.uk/s/FBSPI22/?pageurl=interstitial\">feedback</a> will help us to improve it.",
-  "bullets": [
-    "Hearings in Civil and Family Courts in Milton Keynes, Oxford, and Reading",
-    "Single Justice Procedure cases, including TV licensing and minor traffic offences such as speeding"
-  ],
-  "moreMessage": "More courts and tribunals will become available over time.",
-  "legalMessage": "Legal and media professionals can",
-  "signInLink": "sign in",
-  "welshMessage": "This service is also available in",
-  "languageLink": "Welsh (Cymraeg)",
-  "continue": "Continue",
-  "beforeMessage": "Before you start",
-  "scoAndNI": "If you're in Scotland or Northern Ireland",
-  "contactMessage": "Contact the:",
-  "bulletOptions": [
-    {
-      "linkMessage": "Scottish Courts website",
-      "message": "for courts and some tribunals in Scotland",
-      "link": "https://www.scotcourts.gov.uk/"
-    },
-    {
-      "linkMessage": "Northern Ireland Courts and Tribunals Service",
-      "message": "for courts and tribunals in Northern Ireland",
-      "link": "https://www.courtsni.gov.uk/en-GB/ContactDetails/Pages/default.aspx"
-    }
-  ]
-}
