+++ src/test/routes/home.test.ts
@@ -1,15 +1,23 @@
 import { expect } from 'chai';
+import { app } from '../../main/app';
 import request from 'supertest';
 
-import { app } from '../../main/app';
-
-// TODO: replace this sample test with proper route tests for your application
 describe('Home page', () => {
-  describe('on GET', () => {
-    test('should return sample home page', async () => {
-      await request(app)
-        .get('/')
-        .expect((res) => expect(res.status).to.equal(200));
-    });
+  test('should render a page', async () => {
+    await request(app)
+      .get('/')
+      .expect((res) => expect(res.status).to.equal(200));
+  });
+
+  test('should render a page with a language change to English', async () => {
+    await request(app)
+      .get('/?lng=en')
+      .expect((res) => expect(res.status).to.equal(200));
+  });
+
+  test('should render a page with a language change to Welsh', async () => {
+    await request(app)
+      .get('/?lng=cy')
+      .expect((res) => expect(res.status).to.equal(200));
   });
 });
