--- src/test/routes/interstitial.test.ts
@@ -1,23 +1,0 @@
-import { expect } from 'chai';
-import { app } from '../../main/app';
-import request from 'supertest';
-
-describe('Interstitial page', () => {
-  test('should render a page', async () => {
-    await request(app)
-      .get('/interstitial')
-      .expect((res) => expect(res.status).to.equal(200));
-  });
-
-  test('should render a page with a language change to English', async () => {
-    await request(app)
-      .get('/interstitial?lng=en')
-      .expect((res) => expect(res.status).to.equal(200));
-  });
-
-  test('should render a page with a language change to Welsh', async () => {
-    await request(app)
-      .get('/interstitial?lng=cy')
-      .expect((res) => expect(res.status).to.equal(200));
-  });
-});
